# [PERF] Make the index correctly use FTS